Filter section component: Set `ga4_index` internally
This component already has access to the section index and count,
there's no reason we need to explicitly set this data attribute in every
use of the component.

Also removes an erroneous conversion of 0-based to 1-based indexing (by
adding 1), which we should do consistently in a single place in a future
commit, and removes unnecessary explicit variable use in hash syntax.